+++ assets/themes/ccc45/layout/layout.html
@@ -13,5 +13,7 @@
 
 {% include 'partials/fooetr.html' %}
 
+{% block script %}{% endblock %}
+
 </body>
 </html>
