+++ lib/acyort.js
@@ -7,6 +7,7 @@ var serialize = require('./serialize');
 var templates = require('./template')();
 
 var colors = require('colors');
+var moment = require('moment');
 
 module.exports = function() {
 
@@ -22,27 +23,32 @@ module.exports = function() {
 
     }
 
-
     function build(data) {
 
-        console.log(data.helper)
+        // helper function
+        var helper = {
+            time: function(time, format) {
+                return moment(time).format(format)
+            },
+            post: function(id) {
+                return data.posts.filter(function(post) {
+                    return post.id == id
+                })[0]
+            }
+        }
 
         // page posts
-        /*
         data.pages.forEach(function(page) {
             render(page.path, templates.page, {page: page, helper: helper, data: data})
         })
-        */        
 
         // archives
-        //render('/archives/index.html', templates.archives, {archives: data.archives, helper: helper, data: data})
+        render('/archives/index.html', templates.archives, {archives: data.archives, helper: helper, data: data})
 
         // posts
-        /*
-        for (var id in data.posts) {
-            render(data.posts[id].path, templates.post, {post: data.posts[id], helper})
-        }
-        */
+        data.posts.forEach(function(post) {
+            render(post.path, templates.post, {post: post, helper: helper, data: data})
+        })
 
         /*
         // copy assets
