+++ lib/serialize/index.js
@@ -11,8 +11,6 @@ var _archives = require('./archives');
 var _pager = require('./pager.js');
 var config = require('../config');
 
-var moment = require('moment');
-
 module.exports = function(issues) {
 
     var posts = [];
@@ -142,32 +140,14 @@ module.exports = function(issues) {
         pager.categories[category.name] = _pager(category.posts, 'category/'+ category.name)
     })
 
-    // helper function
-    var helper = {
-        time: function(time, format) {
-            return moment(time).format(format)
-        },
-        post: function(id) {
-            return data.posts.filter(function(post) {
-                return post.id == id
-            })[0]
-        }
-    }
-
     return {
-
-        helper: helper,
-
-        contents: {
-            config: config,
-            posts: posts,
-            pages: pages,
-            categories: categories,
-            tags: tags,
-            archives: archives,
-            pager: pager
-        }
-
+        config: config,
+        posts: posts,
+        pages: pages,
+        categories: categories,
+        tags: tags,
+        archives: archives,
+        pager: pager
     }
 
 }
